Fix: Rspec problems
